Fixed undefined reference to statis method in FileIO
