+++ activation/tanh.hpp
@@ -67,4 +67,3 @@ template class Tanh<double>;
 template class Tanh<float>;
 #endif /* TANH_HPP_ */
 
-
