--- gnunet.cpp
@@ -1,51 +1,0 @@
-/*
- * gnunet.cpp
- *
- *  Created on : Mar 19, 2016
- *      Author : Omar Makke (O jMakke)
- *      Email  : ojmakke@yahoo.com
-
-This file is part of GNU Nets also known as GNUNets
-
-GNU Nets is free software: you can redistribute it and/or modify
-it under the terms of the Affero GNU General Public License as published by
-the Free Software Foundation, version 3 of the License.
-
-GNU Nets is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the Affero
-GNU General Public License for more details.
-
-You should have received a copy of the Affero GNU General Public License
-along with GNU Nets.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include <string.h>
-#include <stdlib.h>
-#include <iostream>
-
-#include "activation/tanh.hpp"
-#include "activation/logistic.hpp"
-#include "activation/step.hpp"
-#include "activation/rectify.hpp"
-#include "helper/nnhelper.hpp"
-#include "helper/console_printer.h"
-#include "helper/fileio.h"
-#include "networks/fullhidden.h"
-#include "parts/edge.h"
-#include "helper/workspace.h"
-
-#define CONTROL 0
-#define INPUT 1
-
-NNHelper<double> nnhelper;
-extern void run_tests();
-
-void switch_state(unsigned *state);
-
-int main(int argc, char* argv[])
-{
-  Workspace::instance().start();
-}
-
-
