+++ helper/randomizer.h
@@ -37,4 +37,3 @@ public:
 
 
 
-
